# Conflicts:
#	core/fileformats/gcodeLoader.cpp
#	core/geometry/raytrace/perlinNoise.h
#	core/reflection/extendedVisitor.cpp
#	core/xml/generated/generated.pri
#	utils/corestructs/coreWidgets/coreWidgets.pri
#	utils/filters/ui/filterWidgets.pri

+++ core/geometry/axisAlignedBox.h
@@ -171,7 +171,7 @@ public:
         vector<Vector3dd> result;
         result.resize(8);
         getPoints(result.data());
-        for (int i = 0; i < result.size(); i++)
+        for (size_t i = 0; i < result.size(); i++)
         {
             result[i] = transform * result[i];
         }
